+++ projects/cm_mcu/commands/I2CCommands.h
@@ -7,9 +7,17 @@
 
 #include <parameters.h>
 
-static BaseType_t i2c_ctl_set_dev(int argc, char **argv);
-static BaseType_t i2c_ctl_r(int argc, char **argv);
-static BaseType_t i2c_ctl_reg_r(int argc, char **argv);
-static BaseType_t i2c_ctl_reg_w(int argc, char **argv);
-static BaseType_t i2c_ctl_w(int argc, char **argv);
-static BaseType_t i2c_scan(int argc, char **argv);
+#ifndef I2C_COMMANDS_H_
+#define I2C_COMMANDS_H_
+
+#define I2C_CTL_MAX_BYTES 4
+
+static BaseType_t i2c_ctl_set_dev(int argc, char **argv, char m);
+static BaseType_t i2c_ctl_r(int argc, char **argv, char m);
+static BaseType_t i2c_ctl_reg_r(int argc, char **argv, char m);
+static BaseType_t i2c_ctl_reg_w(int argc, char **argv, char m);
+static BaseType_t i2c_ctl_w(int argc, char **argv, char m);
+static BaseType_t i2c_scan(int argc, char **argv, char m);
+static BaseType_t i2c_scan(int argc, char **argv, char m);
+
+#endif
