+++ projects/cm_mcu/commands/SensorControl.h
@@ -7,30 +7,29 @@
 
 #include <parameters.h>
 
-static BaseType_t sensor_summary(int argc, char **argv);
+#ifndef SENSOR_CONTROL_H_
+#define SENSOR_CONTROL_H_
+
+static BaseType_t sensor_summary(int argc, char **argv, char m);
 
 // Power
-static BaseType_t power_ctl(int argc, char **argv);
+static BaseType_t power_ctl(int argc, char **argv, char m);
 
 // Alarms
-static BaseType_t alarm_ctl(int argc, char **argv);
+static BaseType_t alarm_ctl(int argc, char **argv, char m);
 
 // LED
-static BaseType_t led_ctl(int argc, char **argv);
+static BaseType_t led_ctl(int argc, char **argv, char m);
 
 // ADC
-static BaseType_t adc_ctl(int argc, char **argv);
-
-// Bootloader
-static BaseType_t bl_ctl(int argc, char **argv);
-
-// Clock
-static BaseType_t clock_ctl(int argc, char **argv);
+static BaseType_t adc_ctl(int argc, char **argv, char m);
 
 // Fireflies
-static BaseType_t ff_ctl(int argc, char **argv);
+static BaseType_t ff_ctl(int argc, char **argv, char m);
 static BaseType_t ff_status(int argc, char **argv);
 
 // FPGA
-static BaseType_t fpga_ctl(int argc, char **argv);
-static BaseType_t fpga_reset(int argc, char **argv);
+static BaseType_t fpga_ctl(int argc, char **argv, char m);
+static BaseType_t fpga_reset(int argc, char **argv, char m);
+
+#endif
