+++ projects/cm_mcu/commands/parameters.h
@@ -16,9 +16,51 @@
 #include "driverlib/sysctl.h"
 #include "driverlib/eeprom.h"
 
+// local includes
+#include "common/i2c_reg.h"
+#include "common/uart.h"
+#include "common/power_ctl.h"
+#include "common/pinsel.h"
+#include "common/smbus.h"
+#include "common/utils.h"
+#include "common/microrl.h"
+
+// FreeRTOS includes
+#include "FreeRTOSConfig.h"
+#include "FreeRTOS.h"
+#include "stream_buffer.h"
+#include "queue.h"
+
+// strlen, strtol, and strncpy
+#include <string.h>
+#include <stdlib.h>
+
+// TivaWare includes
+#include "driverlib/uart.h"
+
+#include "MonitorTask.h"
+#include "CommandLineTask.h"
+#include "I2CCommunication.h"
+#include "Tasks.h"
+#include "AlarmUtilities.h"
+
+#include "clocksynth.h"
+
 int snprintf(char *buf, unsigned int count, const char *format, ...);
 
 #define MAX_INPUT_LENGTH  50
 #define MAX_OUTPUT_LENGTH 512
+#define SCRATCH_SIZE 512
 
-
+extern tSMBus g_sMaster1;
+extern tSMBusStatus eStatus1;
+extern tSMBus g_sMaster2;
+extern tSMBusStatus eStatus2;
+extern tSMBus g_sMaster3;
+extern tSMBusStatus eStatus3;
+extern tSMBus g_sMaster4;
+extern tSMBusStatus eStatus4;
+extern tSMBus g_sMaster6;
+extern tSMBusStatus eStatus6;
+static tSMBus *p_sMaster = &g_sMaster4;
+static tSMBusStatus *p_eStatus = &eStatus4;
